refactor structure

+++ scripts/haddock/run_haddock.sh
@@ -20,13 +20,23 @@ fi
 
 source "$PYEDNA_HOME/config.sh"
 
-RUN_DIR="haddock/run"
-rm -rf "$RUN_DIR"
+SKIP_HADDOCK=true
 
-echo "Starting HADDOCK..."
-conda run -n haddock haddock3 docking_config.cfg
+RUN_DIR="haddock/run"
+
+if [[ "$SKIP_HADDOCK" == false ]]; then
+    rm -rf "$RUN_DIR"
+
+    echo "Starting HADDOCK..."
+    conda run -n haddock haddock3 docking_config.cfg
+else
+    if [[ ! -d "$RUN_DIR" ]]; then
+        echo "Error: SKIP_HADDOCK=true but $RUN_DIR does not exist."
+        exit 1
+    fi
+
+    echo "Using existing HADDOCK run in $RUN_DIR."
+fi
 
 echo "Selecting and processing HADDOCK structures..."
 python -m postprocess_structure
-
-echo "Structure generation complete."
